Trying to add pop-up forms for adding new dates, etc.
Currently have empty, template forms sent to the page; these are copied on
demand for the user.

Next steps are to clean up the design so multiple enter-X forms can
be made for the various trial_entry forms.

Then, to use jQuery.post() to send these enter-X forms over to the
proper /add/X urls.